// src/enums/locationType.ts
export const enum LocationType {
    PUBLIC_LAND = "PUBLIC_LAND",
    PARK = "PARK",
    TRAFFIC_SAFETY_CORRIDOR = "TRAFFIC_SAFETY_CORRIDOR",
    PRIVATE_LAND = "PRIVATE_LAND",
    INDIVIDUAL_HOUSE = "INDIVIDUAL_HOUSE",
    COMMERCIAL_CENTER = "COMMERCIAL_CENTER",
    MARKET = "MARKET",
    GAS_STATION = "GAS_STATION",
    BUS_STOP =  "BUS_STOP"
}